Refactoring. Factor-out 'FolderMetadata'.
