+++ src/libsync/propagateuploadencrypted.h
@@ -54,6 +54,8 @@ private slots:
     void slotFolderEncryptedMetadataError(const QByteArray& fileId, int httpReturnCode);
     void slotUpdateMetadataSuccess(const QByteArray& fileId);
     void slotUpdateMetadataError(const QByteArray& fileId, int httpReturnCode);
+    void slotFinalize(const QString &path = {}, const QString &filename = {}, quint64 size = -1);
+    void slotFolderUnlocked(const QByteArray &folderId, int httpStatus);
 
 signals:
     // Emmited after the file is encrypted and everythign is setup.
@@ -77,7 +79,7 @@ private:
 
   QByteArray _generatedKey;
   QByteArray _generatedIv;
-  QScopedPointer<FolderMetadata> _metadata;
+  FolderMetadata *_metadata = nullptr;
   EncryptedFile _encryptedFile;
   QString _completeFileName;
 };
